pyproject.toml
[project]
name = "regime-locus-matrix"
version = "0.1.1"
description = "RLM - Regime Locus Matrix trading framework"
authors = [{name = "Darrin Vogeli"}]
requires-python = ">=3.10"

dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "pydantic",
    "pyyaml",
    "python-dotenv",
    "statsmodels",
    "numba",
    "yfinance",
    "pandas-datareader",
    "matplotlib",
    "seaborn",
    "plotly",
    "tabulate",
    "jinja2",
    "hmmlearn>=0.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy"
]
# Optional later upgrades (kept out of core install).
options_data = [
    "polygon-api-client",
    "yahooquery"
]
scaling = [
    "polars",
    "dask",
    "pyarrow"
]
# HMM: hmmlearn is included in core dependencies above.
# IBKR TWS API — equity history in ``rlm.data.ibkr_stocks``; option **combo orders** in ``rlm.execution``.
ibkr = [
    "ibapi"
]
# Parquet data lake + optional higher-level IB helpers (see scripts/fetch_ibkr_stock_parquet.py).
datalake = [
    "pyarrow",
]
# Massive Flat Files (S3 gzip CSV) → Parquet; see scripts/ingest_massive_flatfiles_options.py
flatfiles = [
    "boto3",
    "pyarrow",
]
# Alternate IB stack (standalone example only): scripts/examples/ib_insync_fetch_stock_example.py
ib-insync = [
    "ib_insync",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
